+++ src/machine/machine.test.ts
@@ -11,6 +11,7 @@ import {
   constructSend,
   constructValue,
   defaultC,
+  defaultI,
   defaultT,
   fakeWaiter,
 } from '../interpreters/__tests__/fixtures';
@@ -550,7 +551,9 @@ describe('machine coverage', () => {
                 DELAY: 'inc',
               },
               on: {
-                NEXT: '/working',
+                NEXT: {
+                  target: '/working',
+                },
               },
             },
             working: {
@@ -559,7 +562,7 @@ describe('machine coverage', () => {
                 DELAY2: 'inc2',
               },
               on: {
-                FINISH: '/final',
+                FINISH: { target: '/final' },
               },
               states: {
                 fetch: {
@@ -585,7 +588,7 @@ describe('machine coverage', () => {
                           },
                           target: '/working/fetch/idle',
                         },
-                        catch: '/working/fetch/idle',
+                        catch: { target: '/working/fetch/idle' },
                       },
                     },
                   },
@@ -614,7 +617,7 @@ describe('machine coverage', () => {
                             actions: 'write',
                             target: '/working/ui/idle',
                           },
-                          '/working/ui/idle',
+                          { target: '/working/ui/idle' },
                         ],
                       },
                     },
@@ -631,7 +634,7 @@ describe('machine coverage', () => {
             DELAY: 'inc',
           },
           on: {
-            NEXT: '/working',
+            NEXT: { target: '/working' },
           },
         },
         '/working': {
@@ -640,7 +643,7 @@ describe('machine coverage', () => {
             DELAY2: 'inc2',
           },
           on: {
-            FINISH: '/final',
+            FINISH: { target: '/final' },
           },
           states: {
             fetch: {
@@ -666,7 +669,7 @@ describe('machine coverage', () => {
                       },
                       target: '/working/fetch/idle',
                     },
-                    catch: '/working/fetch/idle',
+                    catch: { target: '/working/fetch/idle' },
                   },
                 },
               },
@@ -695,7 +698,7 @@ describe('machine coverage', () => {
                         actions: 'write',
                         target: '/working/ui/idle',
                       },
-                      '/working/ui/idle',
+                      { target: '/working/ui/idle' },
                     ],
                   },
                 },
@@ -727,7 +730,7 @@ describe('machine coverage', () => {
                   },
                   target: '/working/fetch/idle',
                 },
-                catch: '/working/fetch/idle',
+                catch: { target: '/working/fetch/idle' },
               },
             },
           },
@@ -753,7 +756,7 @@ describe('machine coverage', () => {
               },
               target: '/working/fetch/idle',
             },
-            catch: '/working/fetch/idle',
+            catch: { target: '/working/fetch/idle' },
           },
         },
         '/working/ui': {
@@ -780,7 +783,7 @@ describe('machine coverage', () => {
                     actions: 'write',
                     target: '/working/ui/idle',
                   },
-                  '/working/ui/idle',
+                  { target: '/working/ui/idle' },
                 ],
               },
             },
@@ -809,7 +812,7 @@ describe('machine coverage', () => {
                 actions: 'write',
                 target: '/working/ui/idle',
               },
-              '/working/ui/idle',
+              { target: '/working/ui/idle' },
             ],
           },
         },
@@ -817,7 +820,7 @@ describe('machine coverage', () => {
         '/final': {},
       };
 
-      expect(machine2.preflat).toStrictEqual(expected);
+      expect(machine2.preflat).toEqual(expected);
     });
 
     test('#02 => postConfig', () => {
@@ -830,7 +833,7 @@ describe('machine coverage', () => {
               DELAY: 'inc',
             },
             on: {
-              NEXT: '/working',
+              NEXT: { target: '/working' },
             },
           },
           working: {
@@ -839,7 +842,7 @@ describe('machine coverage', () => {
               DELAY2: 'inc2',
             },
             on: {
-              FINISH: '/final',
+              FINISH: { target: '/final' },
             },
             states: {
               fetch: {
@@ -865,7 +868,7 @@ describe('machine coverage', () => {
                         },
                         target: '/working/fetch/idle',
                       },
-                      catch: '/working/fetch/idle',
+                      catch: { target: '/working/fetch/idle' },
                     },
                   },
                 },
@@ -895,7 +898,7 @@ describe('machine coverage', () => {
                           actions: 'write',
                           target: '/working/ui/idle',
                         },
-                        '/working/ui/idle',
+                        { target: '/working/ui/idle' },
                       ],
                     },
                   },
@@ -923,7 +926,7 @@ describe('machine coverage', () => {
                 DELAY: 'inc',
               },
               on: {
-                NEXT: '/working',
+                NEXT: { target: '/working' },
               },
             },
             working: {
@@ -932,7 +935,7 @@ describe('machine coverage', () => {
                 DELAY2: 'inc2',
               },
               on: {
-                FINISH: '/final',
+                FINISH: { target: '/final' },
               },
               states: {
                 fetch: {
@@ -958,7 +961,7 @@ describe('machine coverage', () => {
                           },
                           target: '/working/fetch/idle',
                         },
-                        catch: '/working/fetch/idle',
+                        catch: { target: '/working/fetch/idle' },
                       },
                     },
                   },
@@ -988,7 +991,7 @@ describe('machine coverage', () => {
                             actions: 'write',
                             target: '/working/ui/idle',
                           },
-                          '/working/ui/idle',
+                          { target: '/working/ui/idle' },
                         ],
                       },
                     },
@@ -1006,7 +1009,7 @@ describe('machine coverage', () => {
             DELAY: 'inc',
           },
           on: {
-            NEXT: '/working',
+            NEXT: { target: '/working' },
           },
         },
         '/working': {
@@ -1015,7 +1018,7 @@ describe('machine coverage', () => {
             DELAY2: 'inc2',
           },
           on: {
-            FINISH: '/final',
+            FINISH: { target: '/final' },
           },
           states: {
             fetch: {
@@ -1041,7 +1044,7 @@ describe('machine coverage', () => {
                       },
                       target: '/working/fetch/idle',
                     },
-                    catch: '/working/fetch/idle',
+                    catch: { target: '/working/fetch/idle' },
                   },
                 },
               },
@@ -1071,7 +1074,7 @@ describe('machine coverage', () => {
                         actions: 'write',
                         target: '/working/ui/idle',
                       },
-                      '/working/ui/idle',
+                      { target: '/working/ui/idle' },
                     ],
                   },
                 },
@@ -1104,7 +1107,7 @@ describe('machine coverage', () => {
                   },
                   target: '/working/fetch/idle',
                 },
-                catch: '/working/fetch/idle',
+                catch: { target: '/working/fetch/idle' },
               },
             },
           },
@@ -1131,7 +1134,7 @@ describe('machine coverage', () => {
               },
               target: '/working/fetch/idle',
             },
-            catch: '/working/fetch/idle',
+            catch: { target: '/working/fetch/idle' },
           },
         },
         '/working/ui': {
@@ -1158,7 +1161,7 @@ describe('machine coverage', () => {
                     actions: 'write',
                     target: '/working/ui/idle',
                   },
-                  '/working/ui/idle',
+                  { target: '/working/ui/idle' },
                 ],
               },
             },
@@ -1188,7 +1191,7 @@ describe('machine coverage', () => {
                 actions: 'write',
                 target: '/working/ui/idle',
               },
-              '/working/ui/idle',
+              { target: '/working/ui/idle' },
             ],
           },
         },
@@ -1208,7 +1211,11 @@ describe('machine coverage', () => {
     const machine = createMachine(
       {
         states: {
-          idle: { on: { NEXT: '/state1' } },
+          idle: {
+            on: {
+              NEXT: {},
+            },
+          },
           state1: {
             activities: { DELAY: 'inc' },
             states: {
@@ -1228,10 +1235,15 @@ describe('machine coverage', () => {
       },
       { ...defaultT, eventsMap: { NEXT: {} } },
       {
-        '/': 'idle',
-        '/state1': 'state11',
-        '/state1/state11': 'state111',
-        '/state1/state11/state111': 'state1111',
+        initials: {
+          '/': 'idle',
+          '/state1': 'state11',
+          '/state1/state11': 'state111',
+          '/state1/state11/state111': 'state1111',
+        },
+        targets: {
+          '/idle.on.NEXT': '/state1',
+        },
       },
     );
 
@@ -1297,7 +1309,7 @@ describe('machine coverage', () => {
           {
             invite: 'For /idle',
             expected: {
-              on: { NEXT: '/state1' },
+              on: { NEXT: { target: '/state1' } },
             },
             parameters: '/idle',
           },
@@ -1335,7 +1347,7 @@ describe('machine coverage', () => {
       );
 
       test('Debug', () => {
-        console.warn(
+        console.log(
           'value',
           JSON.stringify(
             machine.retrieveParentFromInitial('/state1/state11/state111'),
@@ -1361,7 +1373,7 @@ describe('machine coverage', () => {
         },
       },
       defaultT,
-      { '/': 'idle' },
+      defaultI,
     );
 
     const service = interpret(machineT, defaultC);